+++ satella/exceptions.py
@@ -123,7 +123,7 @@ class CodedCustomException(CustomException, metaclass=CodedCustomExceptionMetacl
     >>> assert not isinstance(a, MyCode5Diff)
     """
 
-    def __init__(self, message, code=None, *args, **kwargs):
+    def __init__(self, message='', code=None, *args, **kwargs):
         super().__init__(message, code, *args, **kwargs)
         self.message = message  # type: str
         if code is not None: