+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # v2.14.31
 
-
+* added `sleep_interval` to `hang_until_sig`