+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.14.32
 
+* **bugfix release**: change in `read_lines`