+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-# v2.14.32
-
-* **bugfix release**: change in `read_lines`
+# v2.14.33